Only authorized user can edit user profile

+++ app/Http/Controllers/UsersController.php
@@ -11,9 +11,9 @@ class UsersController extends Controller
 {
     public function __construct()
     {
-//        return $this->middleware('auth')->except([
-//
-//        ]);
+        return $this->middleware('auth')->except([
+
+        ]);
     }
 
     /*
@@ -29,6 +29,8 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        //授权认证
+        $this->authorize('update', $user);
         return view('users.edit', compact('user'));
     }
 
@@ -37,23 +39,25 @@ class UsersController extends Controller
      */
     public function update(UserRequest $request, ImageUploadHandler $uploader, User $user)
     {
+        //授权认证
+        $this->authorize('update', $user);
 
         $data = $request->all();
-        $old = $user->avatar;   //旧图片
+        $old = $user->avatar;   //旧图片, 用于修改后删除原来的图片
 
         if ($request->avatar) {
             $result = $uploader->save($request->avatar, 'avatar', $user->id, 333);
             if ($result) {
                 $data['avatar'] = $result['path'];
+
+                //删除旧的图片
+                $img = explode(config('app.url'), $old)[1];
+                $old_img_real_path = public_path() . $img;
+                exec('rm -f ' . $old_img_real_path);
             }
         }
 
         $user->update($data);
-
-        //删除旧的图片
-        $img = explode(config('app.url'), $old)[1];
-        $old_img_real_path = public_path() . $img;
-        exec('rm -f ' . $old_img_real_path);
 
         return redirect()->route('users.show', $user->id)->with('success', '更新个人信息成功');
     }
